moved func for replaceWord to helpers file
